C++11 flags, header files renamed
