+++ NAMESPACE
@@ -19,8 +19,6 @@ exportClasses(SCFind)
 importFrom(DT,dataTableOutput)
 importFrom(DT,datatable)
 importFrom(DT,renderDataTable)
-importFrom(Rcpp,cpp_object_initializer)
-importFrom(Rcpp,evalCpp)
 importFrom(Rcpp,loadModule)
 importFrom(Rcpp,sourceCpp)
 importFrom(SingleCellExperiment,SingleCellExperiment)
@@ -78,5 +76,4 @@ importFrom(stats,aggregate)
 importFrom(stats,p.adjust)
 importFrom(stats,phyper)
 importFrom(stats,setNames)
-useDynLib(scfind , .registration = TRUE)
 useDynLib(scfind)
