+++ R/scfind-package.R
