+++ man/SCFind-class.Rd
