+++ man/buildCellTypeIndex.Rd
