+++ man/findCellTypes.Rd
