+++ man/loadObject.Rd
