+++ man/mergeIndices.Rd
