+++ man/mergeSCE.Rd
