+++ man/queryGene.Rd
