+++ man/scfind.Rd
