+++ man/scfindShinyServer.Rd
