+++ src/Makevars
@@ -1,5 +1,4 @@
-PKG_LIBS = $(LAPACK_LIBS) $(BLAS_LIBS)
-PKG_CXXFLAGS = -std=c++11
+PKG_LIBS = $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
 
 MAKEFLAGS = -j4
